+++ tests/atest/transformers/AlignSettingsSection/test_transformer.py
@@ -1,35 +1,29 @@
 import pytest
 
-from .. import run_tidy_and_compare
+from .. import TransformerAcceptanceTest
 
 
-class TestAlignSettingsSection:
+class TestAlignSettingsSection(TransformerAcceptanceTest):
     TRANSFORMER_NAME = "AlignSettingsSection"
 
     def test_align(self):
-        run_tidy_and_compare(self.TRANSFORMER_NAME, source="test.robot")
+        self.compare(source="test.robot")
 
     def test_align_all_columns(self):
-        run_tidy_and_compare(
-            self.TRANSFORMER_NAME, source="test.robot", expected="all_columns.robot", config=":up_to_column=0"
-        )
+        self.compare(source="test.robot", expected="all_columns.robot", config=":up_to_column=0")
 
     def test_align_three_columns(self):
-        run_tidy_and_compare(
-            self.TRANSFORMER_NAME, source="test.robot", expected="three_columns.robot", config=":up_to_column=3"
-        )
+        self.compare(source="test.robot", expected="three_columns.robot", config=":up_to_column=3")
 
     def test_align_selected_whole(self):
-        run_tidy_and_compare(
-            self.TRANSFORMER_NAME,
+        self.compare(
             source="test.robot",
             expected="selected_whole.robot",
             config=" --startline 1 --endline 25",
         )
 
     def test_align_selected_part(self):
-        run_tidy_and_compare(
-            self.TRANSFORMER_NAME,
+        self.compare(
             source="test.robot",
             expected="selected_part.robot",
             config=" --startline 9 --endline 14",
@@ -37,14 +31,13 @@ class TestAlignSettingsSection:
 
     def test_empty_lines_inside_statement(self):
         # bug from #75
-        run_tidy_and_compare(self.TRANSFORMER_NAME, source="empty_lines.robot")
+        self.compare(source="empty_lines.robot")
 
     def test_continued_statement_style(self):
-        run_tidy_and_compare(self.TRANSFORMER_NAME, source="multiline_keywords.robot")
+        self.compare(source="multiline_keywords.robot")
 
     def test_continued_statement_style_all_columns(self):
-        run_tidy_and_compare(
-            self.TRANSFORMER_NAME,
+        self.compare(
             source="multiline_keywords.robot",
             expected="multiline_keywords_all_col.robot",
             config=":up_to_column=3",
@@ -52,27 +45,23 @@ class TestAlignSettingsSection:
 
     @pytest.mark.parametrize("indent", (0, 2, 20))
     def test_continued_statement_style_all_columns_configure_indent(self, indent):
-        run_tidy_and_compare(
-            self.TRANSFORMER_NAME,
+        self.compare(
             source="multiline_keywords.robot",
             expected=f"multiline_keywords_{indent}indent.robot",
             config=f":up_to_column=3:argument_indent={indent}",
         )
 
     def test_multiline_with_blank_line(self):
-        run_tidy_and_compare(self.TRANSFORMER_NAME, source="blank_line_doc.robot")
+        self.compare(source="blank_line_doc.robot")
 
     def test_doc_multiline_and_whitespace(self):
-        run_tidy_and_compare(self.TRANSFORMER_NAME, source="blank_line_and_whitespace.robot")
+        self.compare(source="blank_line_and_whitespace.robot")
 
     def test_fixed_test(self):
-        run_tidy_and_compare(
-            self.TRANSFORMER_NAME, source="test.robot", expected="test_fixed.robot", config=":min_width=35"
-        )
+        self.compare(source="test.robot", expected="test_fixed.robot", config=":min_width=35")
 
     def test_fixed_all_columns(self):
-        run_tidy_and_compare(
-            self.TRANSFORMER_NAME,
+        self.compare(
             source="test.robot",
             expected="all_columns_fixed.robot",
             config=":min_width=20:up_to_column=0",